Changed some variables to long / size_t to avoid armageddon in some cases

+++ src/compression/lz77.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-uint8_t* lz77_encode(char* txt, int txtlen, int Ls, int Ll, uint32_t* code_length);
-void lz77_decode(uint8_t *code, int codelen, int Ls, int Ll, char* txt);
+uint8_t* lz77_encode(char* txt, size_t txtlen, int Ls, int Ll, size_t* code_length);
+void lz77_decode(uint8_t *code, size_t codelen, int Ls, int Ll, char* txt);
 
 #endif
